refactor
Verify user ownership for offer
