+++ src/core/cli-command/generate-command.ts
@@ -1,7 +1,7 @@
 import got from 'got';
 
-import { MockData } from '../../types/mock-data.js';
 import { CliCommandInterface } from './cli-command.interface.js';
+import { MockData } from '../../types/mock-data.js';
 import OfferGenerator from '../../modules/offer-generator/offer-generator.js';
 import TSVFileWriter from '../file-writer/tsv-file-writer.js';
