+++ src/modules/offer/offer.http
@@ -9,7 +9,7 @@ GET http://localhost:4000/offers/648eec37ce55121b171f3d57 HTTP/1.1
 
 POST http://localhost:4000/offers/ HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImZlbGl1Z2luQG1haWwubmV0IiwiaWQiOiI2NDhlZDE2NjNjYzFkNDAwZjZhMjU5MGEiLCJpYXQiOjE2ODcwODE4MTMsImV4cCI6MTY4NzI1NDYxM30.IXJaOvwzPsgQ5UwW9Roq2cSS7uY66Zo3q57QS7hU08o
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImZlbGl1Z2luQG1haWwubmV0IiwiaWQiOiI2NDhlZDE2NjNjYzFkNDAwZjZhMjU5MGEiLCJpYXQiOjE2ODczNTY3ODgsImV4cCI6MTY4NzUyOTU4OH0.XL98vqazdSd7s72JhYwcJq0f5XTWIvjwR8uVrGZdtAY
 
 {
   "title": "Роскошный отель в Брюсселе",
